Move to Kubernetes.Fluent namespace
